移除仓储层，重写业务层数据库相关代码

+++ src/Good.Admin.Entity/Base_Manage/Base_Action.cs
@@ -40,7 +40,7 @@ namespace Good.Admin.Entity
         /// Default:0
         /// Nullable:False
         /// </summary>           
-        public byte Deleted { get; set; }
+        public int Deleted { get; set; }
 
         /// <summary>
         /// Desc:父级Id
